Adding JetBrains Gradle Changelog plugin - https://github.com/JetBrains/gradle-changelog-plugin
NOTE: This fails when trying to `import org.jetbrains.changelog.date`, so instead I am using `new Date().format('yyyy-MM-dd')` instead of `date()` as the plugin docs suggest. This is hopefully a short workaround while I figure out what is causing this to fail - I'll be opening an issue on GH shortly.